Rename Collaborators Controller
